feat(libprovekit): BindKit wraps binder machinery as a registered Kit (closes #1037)
Sibling to LiftKit (#1033, PR #1036) and LowerKit (#1027, PR #1043). Adapts
the existing binder + named-term-document machinery as a `Kit::transform(&Input)
-> DomainClaim` registered with the public `libprovekit::core::path_executor::KitRegistry`.

## Substrate change

- New `bind.rs` module in `libprovekit::core` exports `BindKit` + the
  per-operation `NamedTerm`/`NamedTermTree`/`NamedWitness`/`NamedTermDocument`
  payload types + the `bind_term_document` entry point + `named_term_document_cid`.
- `BindKit::transform(Input::Term)` walks the term tree, names each operation
  node using the loaded catalog, and emits a `DomainClaim` carrying the
  `NamedTermDocument` as payload.
- Returns `DomainClaim` citing `from` (input Term CID), `to` (named-term-document
  CID), `premises` (prior step's claim CID when consumed as `Input::Claim` from
  a Path step), and the named-term payload.
- `BindError` covers catalog-load failure + parse failure with structured
  refusal info.

## CLI deletion (per hard-refuse "registry without teeth" rule)

- `cmd_bind::run` adapter shim retained for CLI argv parsing.
- Direct binder dispatch removed: `cmd_bind.rs` shrinks from ~1100 lines to
  ~100 (-1024 lines / +43 lines). Production flow is now: argv parse →
  BindKit registered into KitRegistry → `execute_path` → render result.
- No parallel-but-deprecated paths.

## Substrate addition: premise chain propagation

`execute_path` previously executed each step's claim in isolation. BindKit's
multi-step assertion (lift then bind, where bind's claim must cite lift's
output CID in `claim.premises`) requires the executor to walk
`step.depends_on` and extend each step's claim premises with the
prior steps' claim CIDs. Added `step_premises` helper + one line in the
verb-dispatch arm (`claim.premises.extend(step_premises(step, &claims_by_step)?)`).

The change is conservative: empty `depends_on` produces empty premises, so
single-step paths are byte-identical to current behavior.

## Boy-scout: #1044 verb-field oversight

PR #1044 (verb-selector) added `verb: Verb` as a required field on
`PathAlgebra` but missed four construction sites:

- `cmd_transport.rs:1468`
- `cmd_lower.rs:347`
- `cmd_bind_migrate.rs:811`
- `tests/lower_kit_path_integration.rs:188`

Each gets `verb: Verb::Transform,` + a `Verb` import. Without this, `cargo
build -p provekit-cli` fails to compile on current main. CI on the #1044
merge commit has been "in progress" for hours; the gap would surface there
imminently. Fix bundled here under boy-scout-scoping (the work is gated on
main compiling).

## Tests added

- `BindKit::transform` unit tests for `from`/`to`/`premises` CID citation.
- `bind_kit_path_integration`: 3-step path (lift → bind → terminal) asserts
  bind step's `claim.premises` contains the lift step's claim CID.
- `bind_kit_path_integration::lift_then_bind_path_carries_lift_output_and_claim_premise`:
  the binding-assertion test that exercises the new `step_premises`
  premise-extension logic.

## Gates

- `cargo test -p libprovekit`: 20 passed across 5 test groups.
- `cargo test -p provekit-cli`: all groups passed (10 result lines, longest
  test = `bind_migrates_better_sqlite3_to_python_with_cross_language_witness_pairs`
  at 174s).
- Em-dash sweep on the 10 dirty files: zero.
- `git diff --check`: clean.

## Non-goals respected

- No new `provekit-bind-*` crate. Extended existing libprovekit modules.
- No parallel registry. Used `KitRegistry` from #1036.
- No parallel executor. Registered with `execute_path` from #1036.
- No new memento family.
- No body-template catalog data changes.
- No broadening of Path semantics beyond #1025/#1026/#1033/#1027/#1038.

## Forward-compat with #1040

This PR's `register("bind-default", BindKit::new(...))` callsite uses the
current two-arg signature. When #1040 (substrate ConformanceDeclaration)
lands, this callsite adds `ConformanceDeclaration::NonCarrier { reason:
"transforms Input::Term to NamedTerm DomainClaim; emits no target source"
}` on rebase. Per #1040's bidirectional sequencing rule, conflict is
mechanical.

Prerequisites: #1033 (PR #1036 merged for path_executor + KitRegistry),
#1027 (PR #1043 merged for LowerKit precedent), #1038 (PR #1044 merged for
verb-selector + Kit::prove default).
