+++ implementations/rust/provekit-cli/src/cmd_transport.rs
@@ -10,7 +10,7 @@ use std::sync::Arc;
 use clap::Parser;
 use libprovekit::core::{
     execute_path, Cid, HashMapInputCatalog, Input, KitRegistry, LowerKit, Path as CorePath,
-    PathAlgebra, Term,
+    PathAlgebra, Term, Verb,
 };
 use libprovekit::desugar::{load_desugaring_rules_from_dir, DesugaringSet};
 use libprovekit::transport::{transport_term, OperationTransport, TermTransport};
@@ -1470,6 +1470,7 @@ fn realize_spec_via_path(
             kit: kit_name.clone(),
             inputs: vec![input_cid],
             depends_on: vec![],
+            verb: Verb::Transform,
         }],
     }));
     let mut registry = KitRegistry::default();
